// packages/common/src/vk-client/enums.ts
export const enum EAudioTupleIndex {
    ID = 0,
    OWNER_ID = 1,
    URL = 2,
    TITLE = 3,
    PERFORMER = 4,
    DURATION = 5,
    ALBUM_ID = 6,
    AUTHOR_LINK = 8,
    LYRICS = 9,
    FLAGS = 10,
    CONTEXT = 11,
    EXTRA = 12,
    HASHES = 13,
    COVER_URL = 14,
    ADS = 15,
    SUBTITLE = 16,
    MAIN_ARTISTS = 17,
    FEAT_ARTISTS = 18,
    ALBUM = 19,
    TRACK_CODE = 20,
    RESTRICTION = 21,
    ALBUM_PART = 22,
    ACCESS_KEY = 24,
    CHART_INFO = 25,
    TRACK_PAGE_ID = 26,
    IS_ORIGINAL_SOUND = 27,
}